Handle files that exist in both base & android
