+++ baseAndroid/src/main/java/com/smartdevicelink/trace/SdlTraceBase.java
@@ -1,0 +1,1 @@
+../../../../../../../base/src/main/java/com/smartdevicelink/trace/SdlTraceBase.java
