+++ javaSE/src/main/java/com/smartdevicelink/transport/TransportManager.java
